requirements.txt
Flask>=3.0
psycopg2-binary>=2.9
fpdf2>=2.7
gunicorn>=22.0
